Dockerizing everything
This is a major update.
For each of the admin, client, and server there is now an associated docker file and set up.

This allows for more straight forward connections between these parts, better separation of concerns, and better security.

The database file and the images will eventually be moved to a database data folder outside the root "Real Fake News" directory.

The nginx in "client" will be connected to the main nginx via the port number and localhost.